Fix size_t / int conversion warnings in qmake
Fixes

  qtbase\qmake\generators\makefiledeps.cpp(985): warning C4267: '+=': conversion from 'size_t' to 'int', possible loss of data
  qtbase\qmake\generators\makefiledeps.cpp(986): warning C4267: '+=': conversion from 'size_t' to 'int', possible loss of data

Pick-to: 6.2

+++ qmake/generators/makefiledeps.cpp
@@ -982,8 +982,8 @@ bool QMakeSourceFileInfo::findMocs(SourceFile *file)
                                     debug_msg(2, "Mocgen: %s:%d Found \"q%s%s\"",
                                               file->file.real().toLatin1().constData(), line_count,
                                               tag, keyword);
-                                    x += strlen(tag);
-                                    x += strlen(keyword);
+                                    x += static_cast<int>(strlen(tag));
+                                    x += static_cast<int>(strlen(keyword));
                                     ignore[interest] = true;
                                 }
                                 ++interest;
